# EndorsementsRetriever perf, hardening and unit test coverage