Per request from @cleemullins, this PR focuses on the `EndorsementsRetriever` class which is responsible for fetching and parsing keys out of OpenID Connect metadata documents to be used while validating JWT tokens.

Hardening wise I added better exception handling and propagation around HTTP level failures. NOTE: network level failures will still result in a raw `Exception` being propagated.

Test Coverage was initiated in this PR which brings coverage for the `EndorsementsRetriever` class from 0% to 100% coverage.

Performance wise the new implementation brings gains of ~35%-52% in CPU and more modest allocation gains of ~10-20% depending on the scenarios.

----

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3110.0
  Job-KGFYSY : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3110.0

Jit=RyuJit  Platform=X86  Runtime=Clr  

```
|                          Method | JsonIndex |       Mean |    Error |    StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
|-------------------------------- |---------- |-----------:|---------:|----------:|-------:|---------:|-------:|----------:|
|           **GetConfigurationAsync** |         **0** |   **737.1 ns** | **14.57 ns** |  **15.59 ns** |   **1.00** |     **0.00** | **0.6380** |   **2.62 KB** |
| GetConfigurationAsync_OPTIMIZED |         0 |   476.3 ns | 12.49 ns |  35.23 ns |   0.65 |     0.05 | 0.5770 |   2.37 KB |
|                                 |           |            |          |           |        |          |        |           |
|           **GetConfigurationAsync** |         **1** | **2,253.2 ns** | **43.99 ns** |  **55.64 ns** |   **1.00** |     **0.00** | **0.8774** |   **3.61 KB** |
| GetConfigurationAsync_OPTIMIZED |         1 | 1,082.4 ns | 19.58 ns |  18.31 ns |   0.48 |     0.01 | 0.6905 |   2.83 KB |
|                                 |           |            |          |           |        |          |        |           |
|           **GetConfigurationAsync** |         **2** | **2,606.1 ns** | **51.52 ns** |  **99.26 ns** |   **1.00** |     **0.00** | **0.9079** |   **3.73 KB** |
| GetConfigurationAsync_OPTIMIZED |         2 | 1,382.6 ns | 26.04 ns |  48.28 ns |   0.53 |     0.03 | 0.7439 |   3.05 KB |
|                                 |           |            |          |           |        |          |        |           |
|           **GetConfigurationAsync** |         **3** | **3,130.5 ns** | **63.21 ns** | **109.04 ns** |   **1.00** |     **0.00** | **0.9346** |   **3.83 KB** |
| GetConfigurationAsync_OPTIMIZED |         3 | 1,636.2 ns | 31.70 ns |  26.47 ns |   0.52 |     0.02 | 0.7725 |   3.17 KB |
|                                 |           |            |          |           |        |          |        |           |
|           **GetConfigurationAsync** |         **4** | **3,572.7 ns** | **71.36 ns** |  **82.17 ns** |   **1.00** |     **0.00** | **0.9842** |   **4.04 KB** |
| GetConfigurationAsync_OPTIMIZED |         4 | 2,080.9 ns | 26.31 ns |  23.32 ns |   0.58 |     0.01 | 0.8240 |   3.39 KB |
|                                 |           |            |          |           |        |          |        |           |
|           **GetConfigurationAsync** |         **5** | **5,675.6 ns** | **75.83 ns** |  **70.93 ns** |   **1.00** |     **0.00** | **1.1749** |   **4.84 KB** |
| GetConfigurationAsync_OPTIMIZED |         5 | 3,279.1 ns | 58.80 ns |  55.00 ns |   0.58 |     0.01 | 0.9193 |   3.78 KB |
|                                 |           |            |          |           |        |          |        |           |
|           **GetConfigurationAsync** |         **6** | **6,987.5 ns** | **86.19 ns** |  **80.62 ns** |   **1.00** |     **0.00** | **1.2512** |   **5.14 KB** |
| GetConfigurationAsync_OPTIMIZED |         6 | 4,407.9 ns | 45.80 ns |  42.84 ns |   0.63 |     0.01 | 1.0071 |   4.15 KB |